Give BuildInvoker the ability to invoke `blaze info` itself.
Make the running of `blaze info` in the invokers lazy, so that the deprecated methods BuildSystemProvider.getSyncBinaryPath() and BuildSystemProvider.getBinaryPath() do not have the side effect of running a `blaze info` command. This laziness can be removed once those deprecated methods are deleted. Also make sure then `blaze info` is not run more than once per BuildInvoker created by memoizing the result.

Create an abstract build invoker to deal with this logic and enable it so be shared between all BuildInvoker implementations.

PiperOrigin-RevId: 440101092